requirements.txt
appdirs==1.4.0
docopt==0.6.2
mysql-connector-python-rf==2.2.2
packaging==16.8
pyparsing==2.1.10
six==1.10.0
